repos/erc20-gated-server-access/backend/requirements.txt

Flask==2.0.2
Flask-CORS==3.0.10
web3==5.24.0
python-dotenv==0.19.2
